Breaking things up in a pretty naive way for future work I don't have time for.  Messing with bitrate and trying to cleanup the multipass temp files
